% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roll_talk.R
\docType{data}
\name{bench_100}
\alias{bench_100}
\title{Benchmark Data for Width of 100}
\format{
A data frame. Rows are evaluation times and columns are the functions.
}
\usage{
bench_100
}
\description{
A data set with the benchmark results for rolling statistics with a \code{width} of 100.
}
\keyword{datasets}
