% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roll_talk.R
\docType{data}
\name{bench_1000}
\alias{bench_1000}
\title{Benchmark Data for Width of 1,000}
\format{
A data frame. Rows are evaluation times and columns are the functions.
}
\usage{
bench_1000
}
\description{
A data set with the benchmark results for rolling statistics with a \code{width} of 1,000.
}
\keyword{datasets}
