% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roll_talk.R
\name{roll_talk}
\alias{roll_talk}
\title{Rolling and Expanding Statistics}
\usage{
roll_talk(output_dir = getwd(), output_file = "roll_talk.pdf", ...)
}
\arguments{
\item{output_dir}{Directory for the rendered PDF file.}

\item{output_file}{Name of the rendered PDF file.}

\item{...}{Additional arguments passed to \code{\link[quarto:quarto_render]{quarto::quarto_render}}.}
}
\description{
A presentation for rolling and expanding statistics of time-series data.
}
\examples{
\dontrun{
# rolling exploratory data analysis
roll_talk()
}
}
