\name{rolltalk-package}
\alias{rolltalk-package}
\alias{rolltalk}
\docType{package}
\title{
  \packageTitle{rolltalk}
}
\description{
  \packageDescription{rolltalk}
}
\details{
  \code{rolltalk} is a package that provides a presentation of rolling and expanding statistics for time-series data.
}
\author{
  \packageAuthor{rolltalk}
}
